Major refactor, move most code to FPSLicenseHelper
